+++ src/Backends/Vulkan/src/pipeline_layout.cpp
@@ -82,7 +82,7 @@ public:
 
         VkPipelineLayoutCreateInfo pipelineLayoutInfo = {};
         pipelineLayoutInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO;
-        pipelineLayoutInfo.setLayoutCount = descriptorSetLayouts.size();
+        pipelineLayoutInfo.setLayoutCount = static_cast<UInt32>(descriptorSetLayouts.size());
         pipelineLayoutInfo.pSetLayouts = descriptorSetLayouts.data();
         pipelineLayoutInfo.pushConstantRangeCount = static_cast<UInt32>(rangeHandles.size());
         pipelineLayoutInfo.pPushConstantRanges = rangeHandles.data();
